refactor: move run package tests to cli package
