+++ packages/cli/package.json
@@ -303,6 +303,7 @@
     "pretest": "lerna run prepack --concurrency 4 && tsc -p test --noEmit",
     "test": "nyc mocha --forbid-only \"test/**/*.unit.test.ts\"",
     "test:acceptance": "mocha --forbid-only \"test/**/*.acceptance.test.ts\" && node ./bin/bats-test-runner",
+    "test:integration": "mocha --forbid-only \"test/**/*.integration.test.ts\"",
     "test:smoke": "mocha --forbid-only \"test/**/smoke.acceptance.test.ts\"",
     "posttest": "yarn lint",
     "version": "oclif readme --multi && git add README.md ../../docs"
