+++ packages/cli/test/integration/logs.integration.test.ts
