+++ packages/cli/test/integration/run/detached.integration.test.ts
