+++ packages/cli/test/unit/commands/console.unit.test.ts
@@ -1,13 +1,15 @@
+/* eslint-disable @typescript-eslint/ban-ts-comment */
 import {expect, test} from '@oclif/test'
 import * as sinon from 'sinon'
 
-import Dyno from '../../../src/lib/dyno'
+import Dyno from '../../../src/lib/run/dyno'
 
 describe('console', () => {
-  let dynoOpts
+  let dynoOpts: { command: any }
 
   test
     .stub(Dyno.prototype, 'start', sinon.stub().callsFake(function () {
+      // @ts-ignore
       dynoOpts = this.opts
       return Promise.resolve()
     }))
