+++ packages/cli/test/unit/commands/run/detached.unit.test.ts
