+++ packages/cli/test/unit/commands/run/index.unit.test.ts
@@ -1,17 +1,19 @@
+/* eslint-disable @typescript-eslint/ban-ts-comment */
 import {expect, test} from '@oclif/test'
 import * as sinon from 'sinon'
 
-import Dyno from '../../../../src/lib/dyno'
+import Dyno from '../../../../src/lib/run/dyno'
 
 describe('run/index', () => {
   describe('runs correct command', () => {
-    let dynoOpts
+    let dynoOpts: { command: any }
     test
       .nock('https://api.heroku.com', api => {
         api.get('/account')
           .reply(200, {})
       })
       .stub(Dyno.prototype, 'start', sinon.stub().callsFake(function () {
+        // @ts-ignore
         dynoOpts = this.opts
         return Promise.resolve()
       }))
