+++ packages/cli/test/unit/lib/run/helpers.unit.test.ts
@@ -1,6 +1,6 @@
 import {expect, test} from '@oclif/test'
 
-import {buildCommand, buildEnvFromFlag} from '../../../src/lib/helpers'
+import {buildCommand, buildEnvFromFlag} from '../../../../src/lib/run/helpers'
 
 describe('helpers.buildCommand()', () => {
   [
